% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_auc_longer.R
\name{calc_auc_longer}
\alias{calc_auc_longer}
\title{Perform ROC Analysis on Long-Format Data}
\usage{
calc_auc_longer(
  data,
  response_col,
  predictor_col,
  value_col,
  youdex = TRUE,
  threshold = NULL,
  sensitivity_threshold = NULL,
  specificity_threshold = NULL,
  digits = 3
)
}
\arguments{
\item{data}{A data frame containing the response variable, predictor variable, and value column.}

\item{response_col}{The name of the column in \code{data} containing the binary response variable (e.g., 0/1 or "Case"/"Control").}

\item{predictor_col}{The name of the column in \code{data} containing the predictor variable names (e.g., gene names or biomarker names).}

\item{value_col}{The name of the column in \code{data} containing the predictor values (e.g., expression levels or scores).}

\item{youdex}{Logical. If \code{TRUE}, the Youden index is calculated and included in the output. Default is \code{TRUE}.}

\item{threshold}{A numeric value specifying a fixed threshold for sensitivity and specificity calculations. Default is \code{NULL}.}

\item{sensitivity_threshold}{A numeric value specifying the minimum sensitivity threshold. Default is \code{NULL}.}

\item{specificity_threshold}{A numeric value specifying the minimum specificity threshold. Default is \code{NULL}.}

\item{digits}{An integer specifying the number of decimal places to round the results. Default is \code{3}.}
}
\value{
A data frame with evaluation metrics, such as AUC, sensitivity, and specificity, calculated for each predictor:
}
\description{
This function calculates Receiver Operating Characteristic (ROC) curves, Area Under the Curve (AUC),
sensitivity, specificity, and other related metrics for long-format data. It is particularly useful
for datasets where multiple predictors are stored in a single column, and their corresponding values
are stored in another column.
}
\examples{
# Load example data
data("example_data_long")
table(example_data_long$ID,example_data_long$Group)

# Perform ROC analysis
# 1.calculate the evaluation metrics using the Youdex index
result <- calc_auc_longer(data = example_data_long,
   response_col = "Group",
   predictor_col = "ID",
   value_col = "Value")
View(result)

# 2.calculate the evaluation metrics using the cutoff of the threshold
result <- calc_auc_longer(data = example_data_long,
   response_col = "Group",
   predictor_col = "ID",
   value_col = "Value",
   threshold = 0.3,)
View(result)

## 3.calculate the evaluation metrics using the cutoff of the sensitivity(\%)
result <- calc_auc_longer(data = example_data_long,
   response_col = "Group",
   predictor_col = "ID",
   value_col = "Value",
   sensitivity_threshold = 60)
View(result)

## 4.uisng the cutoff of  the specificity(\%)
result <- calc_auc_longer(data = example_data_long,
   response_col = "Group",
   predictor_col = "ID",
   value_col = "Value",
   specificity_threshold = 90)
View(result)
}
