% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_auc_wider.R
\name{calc_auc_wider}
\alias{calc_auc_wider}
\title{Perform ROC Analysis on Wider-Format Data}
\usage{
calc_auc_wider(
  data,
  response_col = NULL,
  youdex = TRUE,
  threshold = NULL,
  sensitivity_threshold = NULL,
  specificity_threshold = NULL,
  digits = 3
)
}
\arguments{
\item{data}{A data frame containing the response variable and predictor variables.}

\item{response_col}{The name of the column in \code{data} containing the binary response variable (e.g., 0/1 or "Case"/"Control").}

\item{youdex}{Logical. If \code{TRUE}, the Youden index is calculated and included in the output. Default is \code{TRUE}.}

\item{threshold}{A numeric value specifying a fixed threshold for sensitivity and specificity calculations. Default is \code{NULL}.}

\item{sensitivity_threshold}{A numeric value specifying the minimum sensitivity threshold. Default is \code{NULL}.}

\item{specificity_threshold}{A numeric value specifying the minimum specificity threshold. Default is \code{NULL}.}

\item{digits}{An integer specifying the number of decimal places to round the results. Default is \code{3}.}
}
\value{
A data frame with evaluation metrics, such as AUC, sensitivity, and specificity, calculated for each predictor:
}
\description{
This function calculates Receiver Operating Characteristic (ROC) curves, Area Under the Curve (AUC),
sensitivity, specificity, and other related metrics for wider-format data. In wider-format data,
each predictor is stored in a separate column, and the function evaluates each predictor independently.
}
\examples{
# Load example data
data("example_data_wide")
# Perform ROC analysis
# 1.calculate the evaluation metrics using the Youdex index
result <- calc_auc_wider(data = example_data_wide, response_col = "Group")
View(result)

# 2.calculate the evaluation metrics using the cutoff of the threshold
result <- calc_auc_wider(data = example_data_wide, response_col = "Group",threshold = 0.3)
View(result)

# 3.calculate the evaluation metrics using the cutoff of the sensitivity(\%)
result <- calc_auc_wider(data = example_data_wide, response_col = "Group", sensitivity_threshold = 60)
View(result)

# 4.calculate the evaluation metrics using the cutoff of  the specificity(\%)
result <- calc_auc_wider(data = example_data_wide, response_col = "Group", specificity_threshold = 90)
View(result)
}
