% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiroc_test.R
\name{multiroc_test}
\alias{multiroc_test}
\title{Compare ROC Curves Between Two Datasets for Multiple Predictors}
\usage{
multiroc_test(data_one, data_two, response_col, method = "delong")
}
\arguments{
\item{data_one}{A data frame representing the first dataset (e.g., training set) containing the binary response variable}

\item{data_two}{A data frame representing the second dataset (e.g., validation set) with the same structure as \code{data_one}.}

\item{response_col}{A character string specifying the name of the binary response column in both datasets (e.g., "Group").}

\item{method}{A character string specifying the method for the ROC comparison test. Options are \code{"delong"}, \code{"bootstrap"},
or \code{"venkatraman"}. The first letter is sufficient. Default is \code{"delong"}.}
}
\value{
A data frame containing two columns: \code{Model} (the predictor name) and \code{P_Value} (the p-value of the ROC test).
}
\description{
This function compares the ROC curves of multiple predictors between two datasets by performing a statistical test
(default is the DeLong test). It calculates the ROC curves for each predictor in both data sets and returns a data frame
with the corresponding p-values, indicating whether there is a significant difference between the ROC curves.
}
\details{
The function first checks that both datasets are data frames and contain the specified response column. It then ensures that
the column names of the two datasets match. For each predictor (i.e., each column in \code{data_one} except the response column),
the function calculates ROC curves for both datasets using the \code{pROC::roc()} function and performs a ROC comparison test
using \code{pROC::roc.test()}. The test method can be specified via the \code{method} parameter.
}
\examples{
# Load example data
data("data_one")
data("data_two")
result <- multiroc_test(data_one = data_one,
   data_two = data_two,
   response_col = "Group")
View(result)
}
